+++ json2dart_dbffi/lib/database/base_db_manager.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import 'package:json2dart_dbffi/database/model/column_info.dart';
+import 'package:json2dart_dbffi/database/model/table_info.dart';
 import 'package:path/path.dart';
 import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
@@ -62,11 +64,28 @@ abstract class BaseDbManager {
   ///查询数据库中所有的表
   Future<List<String>> queryTables() async {
     List<Map<String, Object?>> maps =
-    await _database!.rawQuery("select name from sqlite_master where type='table' order by name;");
-    List<String> _tabls = [];
+        await _database!.rawQuery("select name from sqlite_master where type='table' order by name;");
+    List<String> _tables = [];
     for (Map<String, Object?> map in maps) {
-      _tabls.add(map['name'].toString());
+      _tables.add(map['name'].toString());
     }
-    return _tabls;
+    return _tables;
+  }
+
+  ///查询表信息
+  Future<TableInfo?> queryTableInfo(String name) async {
+    List<Map<String, Object?>> maps = await _database!.rawQuery('select * from sqlite_master where name = "$name"');
+    if (maps.isNotEmpty) {
+      TableInfo tableInfo = TableInfo.fromJson(maps[0]);
+      List<Map<String, Object?>> datas = await _database!.rawQuery("pragma table_info ('$name');");
+      List<ColumnInfo> columns = [];
+      for (Map<String, Object?> map in datas) {
+        ColumnInfo info = ColumnInfo.fromJson(map);
+        columns.add(info);
+      }
+      tableInfo.columns = columns;
+      return tableInfo;
+    }
+    return null;
   }
 }